+++ services/backend/order-service/src/main/java/com/veteroch4k/order/model/OrderCreatedEvent.java
@@ -1,7 +1,7 @@
 package com.veteroch4k.order.model;
 
 public record OrderCreatedEvent(
-    Integer orderId,
-    Integer productId,
-    Integer productQuantity
+    Long orderId,
+    Long productId,
+    Long productQuantity
 ) {}
